--- SECURITY.md
@@ -1,3 +1,0 @@
-# Security Policy
-
-Report vulnerabilities by opening a new issue.
