--- app/server/static/css/styles.css
@@ -1,16 +1,0 @@
-/* Body */
-  body {
-    margin: 0;
-    padding: 0;
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    font-family: "Lato", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
-    background-image: url(/textures/sea.gif);
-    background-repeat: repeat;
-  }
-
-/* Main */
-  main {
-
-  }
