--- app/server/static/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Gracidea | Pokémon Live Map</title>
-  <meta name="description" content="Gracidea, a Pokémon live map">
-  <link rel="stylesheet" href="/css/styles.css">
-</head>
-<body>
-
-  <main></main>
-  <script type="module" src="/gracidea.min.js"></script>
-
-</body>
-</html>
