--- client/css/styles.css
@@ -1,383 +1,0 @@
-/* Body */
-  body {
-    margin: 0;
-    padding: 0;
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    font-family: "Lato", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
-  }
-
-/* Main */
-  main {
-    background-color: #24292E;
-  }
-
-  main.ready {
-    background-color: #5068D0;
-  }
-
-  main.ready.no-sea {
-    background-color: #24292E;
-  }
-
-  main .loading {
-    position: fixed;
-    top: 3rem;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: #24292E;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    color: white;
-  }
-
-  main .loading .stated {
-    display: flex;
-    flex-direction: column;
-    font-size: .7rem;
-    list-style-type: none;
-    height: 6.6rem;
-    overflow: hidden;
-    min-width: 50%;
-  }
-
-  main .loading .stated :nth-child(1) {
-    opacity: .7;
-  }
-  main .loading .stated :nth-child(2) {
-    opacity: .6;
-  }
-  main .loading .stated :nth-child(3) {
-    opacity: .5;
-  }
-  main .loading .stated :nth-child(4) {
-    opacity: .4;
-  }
-  main .loading .stated :nth-child(5) {
-    opacity: .3;
-  }
-  main .loading .stated :nth-child(6) {
-    opacity: .2;
-  }
-  main .loading .stated li {
-    opacity: .1;
-  }
-
-/* Header */
-  header {
-    position: fixed;
-    top: 0; left: 0; right: 0;
-    color: white;
-    z-index: 9001;
-    height: 3rem;
-    background-color: #24292E;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    flex-shrink: 0;
-    overflow-x: auto;
-    flex-wrap: nowrap;
-  }
-
-    header > * {
-      flex-basis: 100%;
-    }
-
-      header.hidden {
-        background-color: transparent;
-        pointer-events: none;
-      }
-
-        header.hidden > :not(.title) {
-          display: none;
-        }
-
-        header.hidden > .title {
-          flex-basis: auto;
-        }
-
-  /* Title */
-    header .title {
-      display: flex;
-      align-items: center;
-      height: 100%;
-      font-size: 1.5rem;
-      font-weight: bold;
-      margin: 0 1rem;
-      text-shadow: 1px 1px 2px black;
-      color: inherit;
-      text-decoration: none;
-      font-style: inherit;
-      transition: color .2s ease;
-      cursor: pointer;
-      pointer-events: auto;
-    }
-
-      header .title:hover {
-        color: #004daa;
-      }
-
-  /** Position **/
-    header .position {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      flex-direction: column;
-      padding-left: 1rem;
-      padding-right: 1rem;
-      min-width: 8rem;
-    }
-
-    header .position small {
-      opacity: .5;
-      font-size: .7rem;
-    }
-
-    header .position span, header .position small {
-      white-space: pre;
-    }
-
-  /* Buttons */
-    nav {
-      display: flex;
-      align-items: center;
-      justify-content: end;
-    }
-
-      nav .button {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        padding: .25rem;
-        border-radius: .25rem;
-        margin: 0 1rem;
-        transition: background-color .2s ease;
-      }
-
-      nav .button:hover {
-        cursor: pointer;
-        background-color: white;
-        color: black;
-      }
-
-      nav .button.active {
-        background-color: #3F4448;
-      }
-
-/* World map */
-  map {
-    position: fixed;
-    top: 3rem;
-    left: 0; right: 0;
-    background-color: rgb(63,200,216,.7);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: calc(100vh - 3rem);
-  }
-
-  /* World map wrapper */
-    map .wrapper {
-      position: relative;
-      height: 680px;
-      width: 600px;
-      overflow: auto;
-    }
-
-    /* Map area */
-      map .area {
-        position: absolute;
-      }
-
-      /* Map area image */
-        .area img {
-          image-rendering: pixelated;
-          image-rendering: -moz-crisp-edges;
-          image-rendering: crisp-edges;
-        }
-
-      /* Map area location */
-        .area .location {
-          position: absolute;
-          width: .5rem;
-          height: .5rem;
-          border: 2px solid #515150;
-          border-radius: .25rem;
-          transition: filter .2s;
-          cursor: pointer;
-        }
-
-        .area .location.city {
-          background-color: #FA5263;
-        }
-
-        .area .location.interest {
-          background-color: #71CA92;
-        }
-
-        .area .location:hover {
-          filter: brightness(2);
-        }
-
-        .area .location.disabled {
-          background-color: gray;
-          cursor: not-allowed;
-        }
-
-/* Wiki page on click */
-  .wiki {
-    position: fixed;
-    right: 0;
-    bottom: 0;
-    height: 100%;
-    width: 100%;
-    max-height: 50%;
-    max-width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #24292E;
-    color: white;
-  }
-
-  .wiki .wiki-loading {
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-  }
-
-  .wiki .wiki-loading small {
-    opacity: .5;
-    font-size: .7rem;
-  }
-
-  .wiki iframe {
-    height: 100%;
-    width: 100%;
-  }
-
-  .wiki .close {
-    position: absolute;
-    top: .5rem;
-    right: 1.5rem;
-    background-color: #24292E;
-    z-index: 7777;
-    padding: .25rem .5rem;
-    border-radius: .5rem;
-    cursor: pointer;
-    transition: background-color .2s ease;
-
-  }
-
-  .wiki .close:hover {
-    background-color: white;
-    color: black;
-  }
-
-  @media screen and (min-width: 1100px) {
-    .wiki {
-      top: 3rem;
-      max-height: 100%;
-      max-width: 50%;
-    }
-  }
-
-/* No scrollbars */
-  *::-webkit-scrollbar {
-    display: none;
-  }
-
-/* Transitions */
-  .fade-enter-active, .fade-leave-active {
-    transition: opacity .2s;
-  }
-  .fade-enter, .fade-leave-to {
-    opacity: 0;
-  }
-
-/* Tooltip texts */
-  .tips {
-    position: relative;
-  }
-
-  .tips .tips-text {
-    visibility: hidden;
-    position: absolute;
-    bottom: .5rem;
-    z-index: 1000;
-    font-size: .8rem;
-    background-color: black;
-    color: white;
-    padding: .25rem .5rem;
-    border-radius: .25rem;
-    transform: translateX(-50%);
-    white-space: nowrap;
-  }
-
-  .tips:hover .tips-text {
-    visibility: visible;
-  }
-
-  .tips .tips-text::after {
-    content: " ";
-    position: absolute;
-    top: 100%;
-    left: 50%;
-    border: .25rem  solid black;
-    border-color: black transparent transparent transparent;
-  }
-
-/* Debug */
-  .debug {
-    position: fixed;
-    top: 2rem;
-    right: 0;
-    color: white;
-    background-color: #24292EE0;
-    border-bottom-left-radius: .5rem;
-    padding: 1.5rem .5rem .5rem;
-    display: flex;
-    flex-direction: column;
-    z-index: 8888;
-  }
-
-    .debug .title {
-      font-weight: bold;
-      font-size: 1.1rem;
-    }
-
-    .debug label {
-      width: 100%;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      transition: background-color .2s ease;
-      padding: .25rem;
-      border-radius: .25rem;
-      cursor: pointer;
-    }
-
-    .debug label:hover {
-      background-color: #3F4448;
-    }
-
-    .debug input[type="text"], .debug input[type="number"] {
-      border-radius: .25rem;
-      outline: none;
-      border: none;
-      padding: .125rem .5rem .25rem;
-      background-color: #24292E;
-      width: 4rem;
-      color: white;
-    }
-
-    .debug input[disabled], .debug .disabled {
-      cursor: not-allowed;
-      color: inherit;
-      opacity: .5;
-    }
