+++ source/server/static/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Gracidea | Pokémon Live Map</title>
+  <meta name="description" content="Gracidea, a Pokémon live map">
+  <link rel="stylesheet" href="/css/styles.css">
+</head>
+<body>
+
+  <main></main>
+
+  <div id="location">
+    <div class="position"></div>
+    <div class="name"></div>
+  </div>
+
+  <nav>
+    <button>MAP</button>
+    <button>OPTIONS</button>
+    <button>Test</button>
+    <button>GITHUB</button>
+  </nav>
+
+
+  <script type="module" src="/gracidea.min.js"></script>
+
+</body>
+</html>
